Further signed/unsigned updates.  Fixed other GImage type conversion warnings.

+++ source/G3Dcpp/BinaryInput.cpp
@@ -172,7 +172,7 @@ BinaryInput::~BinaryInput() {
 
 
 void BinaryInput::readBytes(size_t n, void* bytes) {
-    debugAssertM(pos + n <= length, "Read past end of file");
+    debugAssertM((int)(pos + n) <= length, "Read past end of file");
     debugAssert(isValidPointer(bytes));
 
     memcpy(bytes, buffer + pos, n);
@@ -210,7 +210,7 @@ uint64 BinaryInput::readUInt64() {
 
 
 std::string BinaryInput::readString(size_t n) {
-    debugAssertM(pos + n <= length, "Read past end of file");
+    debugAssertM((int)(pos + n) <= length, "Read past end of file");
     
     char *s = (char*)malloc(n + 1);
     assert(s != NULL);
